Refactor differ/callback contract enforcement into a SpreadSheetDiffCallbackBase class that the other callback modules extend.  Remove some unnecessary asstions in UnifiedDiffSpreadSheetDiffCallback.
